+++ chronosphere/intschema/shared_schemas.go
@@ -140,6 +140,7 @@ type PartitionFilterCondition struct {
 	DatasetId    tfid.ID                                `intschema:"dataset_id,optional"`
 	LogFilter    *PartitionFilterConditionLogFilter     `intschema:"log_filter,optional,list_encoded_object"`
 	MetricFilter []PartitionFilterConditionMetricFilter `intschema:"metric_filter,optional"`
+	TraceFilter  *PartitionFilterConditionTraceFilter   `intschema:"trace_filter,optional,list_encoded_object"`
 }
 
 type PartitionFilterConditionLogFilter struct {
@@ -149,6 +150,21 @@ type PartitionFilterConditionLogFilter struct {
 type PartitionFilterConditionMetricFilter struct {
 	Name      string `intschema:"name"`
 	ValueGlob string `intschema:"value_glob"`
+}
+
+type PartitionFilterConditionTraceFilter struct {
+	SpanFilter []PartitionFilterConditionTraceFilterSpanFilter `intschema:"span_filter,optional"`
+}
+
+type PartitionFilterConditionTraceFilterSpanFilter struct {
+	Duration        *TraceDurationFilter `intschema:"duration,optional,list_encoded_object"`
+	Error           *TraceBoolFilter     `intschema:"error,optional,list_encoded_object"`
+	IsRootSpan      *TraceBoolFilter     `intschema:"is_root_span,optional,list_encoded_object"`
+	Operation       *TraceStringFilter   `intschema:"operation,optional,list_encoded_object"`
+	ParentOperation *TraceStringFilter   `intschema:"parent_operation,optional,list_encoded_object"`
+	ParentService   *TraceStringFilter   `intschema:"parent_service,optional,list_encoded_object"`
+	Service         *TraceStringFilter   `intschema:"service,optional,list_encoded_object"`
+	Tag             []TraceTagFilter     `intschema:"tag,optional"`
 }
 
 type RegexParser struct {